첫번째 과제 완료!
(\^ㅁ^/)

+++ src/com/sample/bae/student/FirstPractice.java
@@ -6,7 +6,7 @@ public interface FirstPractice {
 
   void setData(List<String> list);
 
-  Long totalScoreOfMember(String name);
+  Double totalScoreOfMember(String name);
 
   String highestTotalScoreMember();
